hotfix PSync Redirection Data
